compose v3 with custom network settings
I've built on your example today and got it to work. The trick was to set `LIBPROCESS_IP` correctly and since I had to hardcode IPs I thought to use a custom subnet; I'm assuming that this custom subnet won't conflict with existing containers that may run on the host.